// src/app/models/zaposlen.ts
export class Zaposlen {
  _id: string;
  ime: string;
  prezime: string;
  korime: string;
  lozinka: string;
  email: string;
  adresa: string;
  mobilni: string;
  vebsajt: string;
  bio: string;
  tipZvanja: string;
  zvanje: string;
  kabinet: number;
  status: string;
  slikaPath: string;
  lozinkaChanged: boolean;
}